+++ src/main/java/com/epam/esm/dto/CertificateResponseModel.java
@@ -10,10 +10,10 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CertificateResponseModel {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
-    private Integer price;
+    private Long price;
     private Integer duration;
     private String createDate;
     private String lastUpdateDate;
